+++ src/test/java/com/mageddo/dnsproxyserver/docker/ContainerSolvingServiceCompTest.java
@@ -3,8 +3,8 @@ package com.mageddo.dnsproxyserver.docker;
 import com.mageddo.dnsproxyserver.di.Context;
 import com.mageddo.dnsproxyserver.templates.docker.InspectContainerResponseTemplates;
 import com.mageddo.dnsproxyserver.templates.docker.NetworkTemplates;
-import com.mageddo.utils.dagger.mockito.DaggerTest;
-import com.mageddo.utils.dagger.mockito.InjectMock;
+import dagger.sheath.junit.DaggerTest;
+import dagger.sheath.junit.InjectMock;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
