+++ src/test/java/com/mageddo/dnsproxyserver/server/rest/EnvResourceCompTest.java
@@ -3,8 +3,8 @@ package com.mageddo.dnsproxyserver.server.rest;
 import com.mageddo.dnsproxyserver.config.Configs;
 import com.mageddo.dnsproxyserver.config.entrypoint.JsonConfigs;
 import com.mageddo.dnsproxyserver.di.Context;
-import com.mageddo.utils.dagger.Events;
-import com.mageddo.utils.dagger.mockito.DaggerTest;
+import dagger.Events;
+import dagger.sheath.junit.DaggerTest;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
