--- src/test/java/com/mageddo/utils/dagger/mockito/Fields.java
@@ -1,4 +1,0 @@
-package com.mageddo.utils.dagger.mockito;
-
-public class Fields {
-}
