+++ src/test/java/dagger/sheath/junit/CtxWrapper.java
@@ -1,7 +1,6 @@
-package com.mageddo.utils.dagger.mockito;
+package dagger.sheath.junit;
 
-import com.mageddo.utils.dagger.MethodUtils;
-import com.mageddo.utils.dagger.mockito.binding.BindingMethod;
+import dagger.sheath.junit.binding.BindingMethod;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.reflect.FieldUtils;
