+++ src/test/java/dagger/sheath/junit/CtxWrapperTest.java
@@ -1,4 +1,4 @@
-package com.mageddo.utils.dagger.mockito;
+package dagger.sheath.junit;
 
 import org.junit.jupiter.api.Test;
 import sheath.stubing.AppByProvider;
