+++ src/test/java/dagger/sheath/junit/EventHandler.java
@@ -1,4 +1,4 @@
-package com.mageddo.utils.dagger.mockito;
+package dagger.sheath.junit;
 
 public interface EventHandler<T> {
   /**
