+++ src/test/java/dagger/sheath/junit/Fields.java
@@ -1,0 +1,4 @@
+package dagger.sheath.junit;
+
+public class Fields {
+}
