+++ src/test/java/dagger/sheath/junit/NopSupplier.java
@@ -1,4 +1,4 @@
-package com.mageddo.utils.dagger.mockito;
+package dagger.sheath.junit;
 
 import java.util.function.Supplier;
