+++ src/test/java/dagger/sheath/junit/binding/BindingMapMethod.java
@@ -1,6 +1,6 @@
-package com.mageddo.utils.dagger.mockito.binding;
+package dagger.sheath.junit.binding;
 
-import com.mageddo.utils.dagger.mockito.CtxWrapper;
+import dagger.sheath.junit.CtxWrapper;
 
 import javax.inject.Provider;
 import java.lang.reflect.InvocationTargetException;
